+++ qrb_inference_manager/include/qnn_inference/qnn_inference_impl.hpp
@@ -14,6 +14,7 @@
 #include <numeric>
 #include <vector>
 
+#include "HTP/QnnHtpDevice.h"
 #include "HTP/QnnHtpMem.h"
 #include "QnnInterface.h"
 #include "QnnMem.h"